requirements.txt
colorama
pyyaml
requests
shodan
python-owasp-zap-v2.4
python-magic
